nodejs/pyllmapi/requirements.txt
Flask>=2.2.0,<3.0.0
Flask-RESTful==0.3.10
Flask-Cors==5.0.0
gunicorn==20.1.0
#langchain==0.3.0
#langchain-community==0.3.0
langchain-core==0.3.0
langchain-openai==0.2.0
azure-identity==1.18.0
azure-monitor-query==1.4.0
pandas==2.2.3
langchain-community==0.3.0
opencensus-ext-azure==1.1.13
opencensus-ext-flask==0.8.2
applicationinsights==0.11.10
